Improve stunnel example to behave more like a stunnel program.
The client side is behaving like a real stunnel now, waiting local connection
and relaying it through the TLS connection and back to the local connection.

The server side is improved, however it doesn't properly relay it to the local
port on the server. For now it prints the message to stdout and reply a constant
to a client. it waits for EOF from the client before finishing.